Catch IOError exception in wrapper processing script to prevent termination issues
When not catched, any IOError occuring while copying the XSL stylesheet to the
output directory would have made the wrapper processing script fail. As a consequence,
no entrie would be added to the processing task table, even if the job is successful.

If the job was successful but an IOError it not catched, the resulting data would be
copied to the result output directory but nothing would be written to the database(!)