Bug #169: automatic stack file ingestion after successful Swarp processing
The Swarp plugin has been updated (along with the wrapper processing script)
to allow ingestion of the final stack FITS image in the DB.

Policy: during stack ingestion, when a stack with the name already exists, the
current stack is ingested under a new name (old name + '_' + integer). In this case,
the config file (which is saved with the Swarp processing) is modified on-the-fly
to replace the content of the IMAGEOUT value with the new image name. Finally, the
stack image is renamed automatically in the results output directory to match the
new name too.