local_conf.py is now mandatory, but can be empty
The checksetup command looks for the file and creates it if
not found. The settings.py script try to import it unconditionaly.
This is because we must explicitely give the path to local_conf.py
when generating (and executing) all the Condor submission files (CSF).

So this file must be existing, but can be empty.