Integration of django_evolution application
Django Evolution is an extension to Django that allows you to track changes in
your models over time, and to update the database to reflect those changes.

Current merged svn code revision is 166. django_evolution is now distributed with
Youpi so we can fully support it.

It's open source software: http://code.google.com/p/django-evolution/