Condor Python API improved: new CondorQueue class
The CondorQueue class is useful to monitor jobs submitted
to the Condor cluster. It works from the CLI and will be
used by Youpi to improve performance (and reduce code
redundancy) on the 'Active monitoring' page.

For example, it can be used this way:

>>> from terapix.lib.cluster.condor import CondorQueue
>>> q = CondorQueue()
>>> jobs, total = q.getJobs()
>>> jobs[-1]
<CondorJob 1705.32 is idle>
>>> jobs[0]
<CondorJob 1705.17 is running on slot4@xxxxxx since 00:17:35>
>>> jobs[0].isRunning()
True

Some performance gain is achieved by not asking Condor for XML
data at all. This way, responses are faster and lighter.