All plugins: rewrote part dealing with XSL_URL config parameter
Now, when the XSL_URL value uses the file:/// protocol, Youpi copies
the file to the final results output directory. For any other
protocol used to define a target location to use for XSL stylesheet,
the value is passed as is on the command line.

This way, using XSL stylesheets available over HTTP or public FTP
protocols will work fine.