Added support for JsUnit (Javascript unit testing framework)
JsUnit 2.2 (http://www.jsunit.net/) is part of the Youpi distribution.
It can be accessed from the web interface by clicking on the 'Unit Testing'
menu link when in debug mode (DEBUG=True).

A basic test suite has been defined, along with 2 assertion tests in order
to get a fully working, ready-to-use Javascript framework for Youpi.