Bug 198: QFits: fix XML parsing of psfex, compute (and store) image seeing in arcsec
The standalone changeseeing (to be renamed) script is working fine. It
allows to reparse all successful qfits output data and reingest the results
in the database, which may be useful.

The wrapper processing script has been improved. When parsing output of psfex,
any parsing error exception that may be raised in case of old psfex.xml data
format will be catched and a call to the old parsing algorithm for psfex will
be made instead (by calling the parse_psfex_old_xml() function).

The changeseeing could be improved a little bit, since for the moment if forces
to loop over all the successfull qfits. An option will be added to skip a qfits
entry if some values are not NULL.