Running Stiff jobs on the cluster is fully working
The Stiff plugin uses a preprocessing script whose goal is to check
whether the input image is a MEF. In this case, a Swarp is performed
before doing using Stiff.

Note that there are still some issues in the processing cart and the
results page for Stiff jobs is rather incomplete at the moment.