Fixed transaction issue in wrapper processing script
This bug was introduced with changeset 1030:839b81258f87. The
transaction was comitted to the DB just after pre-processing, resulting
in a new failed job entry on the results page when the job gets running
on the cluster. This was not the desired behaviour. Now the transaction is
started at a later time, after the processing is finished, at the beginning of
the post-processing part.