Bug 215: Swarp: add support for automatic compression algo. detection for weight maps
The standalone preprocessing script generated by the Swarp plugin during
CSF generation has been moved to an external template (.tmpl) python file
that will be substituted at runtime.

The preprocessing script locates weight maps in either compressed or
uncompressed form, transfer them locally on the cluster node, and uncompress
them on demand with the right tool (imcopy for .fz files or gunzip for .gz
files).

Note that the condor_tranfert.pl script has been modified with a new -k
option that prevent transferred files to be deleted once the tranfer of
the selection is complete.