Swarp: new final stack ingestion policy!
Until now, final stacks produced by Swarp in Youpi were ingested with
the '_#' policy, that is if a stack with the same name already exists
in the database, the current stack is ingested with a new generated
name which is basename + '_#' (incremental).

This can cause problems if someone wants to (or has to) reprocess Swarp
jobs several times.

The new policy is to always keep the same image name for the final ingested
stack. So, the latest stack ingested overwrites old information.