All plugins: better handling of external XSL stylesheets
In order to get a clean workaround of browsers' cross-sites
scripting security handling, the XSL stylesheets that come
with the low level tool in the suite (except for QFits) are
copied locally by the wrapper processing script.

As a result, the XSL_URL option - that can be user-defined in
all configuration files - is ignored. Youpi forces the value to
be -XSL_URL plugind_id.xsl.

Note also that the '-WRITE_XML Y' option has been hardwritten for
the sextractor, swarp and scamp plugins.