New xslfix.py script to fix XSL paths in XML VOTables
Recursively locates XML VOTABLEs files and replace the path
to the XSL stylesheet by XSL filename in the local directory.
The script automatically copies the matching XSL stylesheet
in the result directy.

$ ./xslfix.py /path/to/swarp

will run a simulation (data not altered). Use the --real option
to really alter the content of the XML and copy the XSL locally.